Clean webpacks and package json
