+++ .github/workflows/cd.yaml
@@ -127,7 +127,6 @@ jobs:
                   echo AWS_DEFAULT_REGION=$AWS_DEFAULT_REGION >> $GITHUB_ENV
 
                   curl -H "Authorization: bearer $ACTIONS_ID_TOKEN_REQUEST_TOKEN" "$ACTIONS_ID_TOKEN_REQUEST_URL" | jq -r '.value' > $AWS_WEB_IDENTITY_TOKEN_FILE
-              if: always()
 
             - name: Run Smoke Test (NPM CJS)
               env:
@@ -138,7 +137,6 @@ jobs:
                   INSTALL_METHOD: 'NPM-CJS'
               run: npm run smoke:headless
               timeout-minutes: 10
-              if: always()
 
             - name: Fetch AWS Credentials for Smoke Test
               run: |
@@ -151,7 +149,6 @@ jobs:
                   echo AWS_DEFAULT_REGION=$AWS_DEFAULT_REGION >> $GITHUB_ENV
 
                   curl -H "Authorization: bearer $ACTIONS_ID_TOKEN_REQUEST_TOKEN" "$ACTIONS_ID_TOKEN_REQUEST_URL" | jq -r '.value' > $AWS_WEB_IDENTITY_TOKEN_FILE
-              if: always()
 
             - name: Run Smoke Test (CDN)
               env:
@@ -161,7 +158,6 @@ jobs:
                   NAME: ${{ secrets.SMOKE_MONITOR_NAME }}
                   INSTALL_METHOD: 'CDN'
               run: npm run smoke:headless
-              if: always()
 
             - name: Publish to NPM
               run: npm publish
