+++ .prettierignore
@@ -6,4 +6,3 @@ src/events
 package-lock.json 
 docs/
 .vscode
-smoke
