+++ smoke/smoke-test-application-NPM-ES/package.json
@@ -1,10 +1,6 @@
 {
     "name": "smoke-test-application-NPM-ES",
     "version": "1.0.0",
-    "npm-pretty-much": {
-        "runRelease": "release",
-        "runTest": "never"
-    },
     "scripts": {
         "build": "webpack --config webpack/webpack.dev.js",
         "release": "npm run build",
@@ -19,20 +15,11 @@
         "babel-loader": "~8.0.6",
         "case-sensitive-paths-webpack-plugin": "^2.1.2",
         "copy-webpack-plugin": "^6.3.2",
-        "http-server": "^13.0.2",
-        "husky": "^7.0.4",
-        "lint-staged": "^12.4.1",
-        "prettier": "2.2.1",
         "ts-loader": "^9.2.6",
-        "tslint": "^5.11.0",
-        "tslint-config-prettier": "^1.18.0",
         "typescript": "^4.4.3",
         "webpack": "^5.58.1",
-        "webpack-bundle-analyzer": "^4.5.0",
-        "webpack-cli": "^4.9.0",
         "webpack-dev-middleware": "^5.2.1",
-        "webpack-dev-server": "^4.3.1",
-        "webpack-merge": "^5.8.0"
+        "webpack-dev-server": "^4.3.1"
     },
     "dependencies": {
         "@babel/runtime": "^7.16.0",
